feat(identity): move auth function to identity-info page and refine ID input handling
- Moved authentication logic directly into `identity-info/page.tsx` for better page-level control
- Updated ID input `onChange` behavior for cleaner and more reliable value handling
